dynap

+++ examples/orbslam_stereo_kittim.py
@@ -40,11 +40,11 @@ def main(vocab_path, settings_path, sequence_path, coco_path, device):
     kernel = cv2.getStructuringElement(cv2.MORPH_ELLIPSE,(2*dilation+1,2*dilation+1))
     for idx in range(num_images):
         left_image = cv2.imread(left_filenames[idx], cv2.IMREAD_UNCHANGED)
-        left_mask = get_mask(coco_demo,left_image)
+        left_mask = get_mask(coco_demo,left_image).astype(np.uint8)
         left_mask_dil = cv2.dilate(left_mask,kernel)
         left_mask -= left_mask_dil
         right_image = cv2.imread(right_filenames[idx], cv2.IMREAD_UNCHANGED)
-        right_mask = get_mask(coco_demo, right_image)
+        right_mask = get_mask(coco_demo, right_image).astype(np.uint8)
         right_mask_dil = cv2.dilate(right_mask, kernel)
         right_mask -= right_mask_dil
         tframe = timestamps[idx]
